tykkäys, security jatkoa

+++ src/main/java/hh/sof3as3/Reseptikirjasto/web/CommentController.java
@@ -5,6 +5,8 @@ import java.sql.Timestamp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +14,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 
 import hh.sof3as3.Reseptikirjasto.domain.Comment;
 import hh.sof3as3.Reseptikirjasto.domain.CommentRepository;
+import hh.sof3as3.Reseptikirjasto.domain.Recipe;
+import hh.sof3as3.Reseptikirjasto.domain.User;
 import hh.sof3as3.Reseptikirjasto.domain.UserRepository;
 
 @Controller
@@ -21,6 +25,13 @@ public class CommentController {
 	@Autowired
 	private CommentRepository commentRepository;
 	@Autowired UserRepository userRepository;
+	
+	// apumetodi kirjautuneen käyttäjän hakemiseen
+	private User getCurrentUser() {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		User user = userRepository.findByUsername(authentication.getName());
+		return user;
+	}
 	
 	// kommentin lisäys
 	@PostMapping("/savecomment")
@@ -33,12 +44,18 @@ public class CommentController {
 	}
 	
 	// kommentin poistaminen
-	// reseptin id redirectiä ja kommentin id kommentin poistamista varten
 	@GetMapping("/deletecomment/{commentid}")
-	@PreAuthorize("hasAnyAuthority('ADMIN')") // todo: admin ja kommentin omistaja
+	@PreAuthorize("hasAnyAuthority('ADMIN', 'USER')")
 	public String deletecomment(@PathVariable("commentid") Long commentid) {
-		Long recipeid = commentRepository.findById(commentid).get().getRecipe().getId();
-		commentRepository.deleteById(commentid);
+		User user = getCurrentUser();
+		// haetaan resepti commentid:n avulla
+		Recipe recipe = commentRepository.findById(commentid).get().getRecipe();
+		// haetaan reseptin id redirectiä varten
+		Long recipeid = recipe.getId();
+		// poisto onnistuu jos kirjautunut käyttäjä on kommentin omistaja tai admin
+		if (recipe.getAuthor() == user || user.getRole() == "ADMIN") {
+			commentRepository.deleteById(commentid);
+		}
 		return "redirect:/recipe/" + recipeid; // uudelleenohjaus takaisin reseptinäkymään
 	}
